refactor: create sui token
